rename upgrade to v6patch1
