+++ app/app.go
@@ -132,7 +132,7 @@ import (
 	"github.com/tokenize-x/tx-chain/v6/app/openapi"
 	appupgrade "github.com/tokenize-x/tx-chain/v6/app/upgrade"
 	appupgradev6 "github.com/tokenize-x/tx-chain/v6/app/upgrade/v6"
-	appupgradev6fixdenommetadata "github.com/tokenize-x/tx-chain/v6/app/upgrade/v6_fix_denom_metadata"
+	appupgradev6patch1 "github.com/tokenize-x/tx-chain/v6/app/upgrade/v6patch1"
 	"github.com/tokenize-x/tx-chain/v6/docs"
 	"github.com/tokenize-x/tx-chain/v6/pkg/config"
 	"github.com/tokenize-x/tx-chain/v6/pkg/config/constant"
@@ -1204,7 +1204,7 @@ func New(
 			app.AccountKeeper.AddressCodec(),
 			app.StakingKeeper.ValidatorAddressCodec(),
 		),
-		appupgradev6fixdenommetadata.New(
+		appupgradev6patch1.New(
 			app.ModuleManager,
 			app.configurator,
 			app.BankKeeper,
